BUGFIX: Fix OptionsSchema check for "callable" options for PHP 8+
Adjust the schema check for `callable` options to only check for the syntax
not whether the method is actually callable.

The behavior of the `is_callable()` function has been changed with PHP 8:
Previously `is_callable('<ClassName>::<methodName>')` returned `true` whether
the method was static or not.
With PHP 8 and newer that check returns `false`.

Fixes: #2